Refactor unzipRemoteFile
